chore: Add OpenAPI support for autotranslate.getSupportedLanguages and autotranslate.saveSettings
