+++ .changeset/openapi-autotranslate-getSupportedLanguages-saveSettings.md
@@ -1,0 +1,6 @@
+---
+'@rocket.chat/meteor': minor
+'@rocket.chat/rest-typings': minor
+---
+
+Add OpenAPI support for the Rocket.Chat autotranslate.getSupportedLanguages and autotranslate.saveSettings API endpoints by migrating to a modern chained route definition syntax and utilizing shared AJV schemas for validation to enhance API documentation and ensure type safety through response validation.
