+++ apps/meteor/app/api/server/v1/autotranslate.ts
@@ -1,12 +1,6 @@
-import type { IMessage } from '@rocket.chat/core-typings';
+import type { IMessage, ISupportedLanguage } from '@rocket.chat/core-typings';
 import { Messages } from '@rocket.chat/models';
-import {
-	ajv,
-	validateUnauthorizedErrorResponse,
-	validateBadRequestErrorResponse,
-	isAutotranslateSaveSettingsParamsPOST,
-	isAutotranslateGetSupportedLanguagesParamsGET,
-} from '@rocket.chat/rest-typings';
+import { ajv, validateUnauthorizedErrorResponse, validateBadRequestErrorResponse } from '@rocket.chat/rest-typings';
 
 import { getSupportedLanguages } from '../../../autotranslate/server/functions/getSupportedLanguages';
 import { saveAutoTranslateSettings } from '../../../autotranslate/server/functions/saveSettings';
@@ -15,14 +9,114 @@ import { settings } from '../../../settings/server';
 import type { ExtractRoutesFromAPI } from '../ApiClass';
 import { API } from '../api';
 
-API.v1.addRoute(
-	'autotranslate.getSupportedLanguages',
-	{
-		authRequired: true,
-		validateParams: isAutotranslateGetSupportedLanguagesParamsGET,
+type AutotranslateGetSupportedLanguagesParamsGET = {
+	targetLanguage: string;
+};
+
+const AutotranslateGetSupportedLanguagesParamsGETSchema = {
+	type: 'object',
+	properties: {
+		targetLanguage: {
+			type: 'string',
+		},
 	},
-	{
-		async get() {
+	required: ['targetLanguage'],
+	additionalProperties: false,
+};
+
+const isAutotranslateGetSupportedLanguagesParamsGET = ajv.compile<AutotranslateGetSupportedLanguagesParamsGET>(
+	AutotranslateGetSupportedLanguagesParamsGETSchema,
+);
+
+type AutotranslateSaveSettingsParamsPOST = {
+	roomId: string;
+	field: 'autoTranslate' | 'autoTranslateLanguage';
+	value: boolean | string;
+	defaultLanguage?: string;
+};
+
+const AutotranslateSaveSettingsParamsPostSchema = {
+	type: 'object',
+	properties: {
+		roomId: {
+			type: 'string',
+		},
+		field: {
+			type: 'string',
+			enum: ['autoTranslate', 'autoTranslateLanguage'],
+		},
+		value: {
+			anyOf: [{ type: 'boolean' }, { type: 'string' }],
+		},
+		defaultLanguage: {
+			type: 'string',
+			nullable: true,
+		},
+	},
+	required: ['roomId', 'field', 'value'],
+	additionalProperties: false,
+};
+
+const isAutotranslateSaveSettingsParamsPOST = ajv.compile<AutotranslateSaveSettingsParamsPOST>(
+	AutotranslateSaveSettingsParamsPostSchema,
+);
+
+type AutotranslateTranslateMessageParamsPOST = {
+	messageId: string;
+	targetLanguage?: string;
+};
+
+const AutotranslateTranslateMessageParamsPostSchema = {
+	type: 'object',
+	properties: {
+		messageId: {
+			type: 'string',
+		},
+		targetLanguage: {
+			type: 'string',
+			nullable: true,
+		},
+	},
+	required: ['messageId'],
+	additionalProperties: false,
+};
+
+const isAutotranslateTranslateMessageParamsPOST = ajv.compile<AutotranslateTranslateMessageParamsPOST>(
+	AutotranslateTranslateMessageParamsPostSchema,
+);
+
+const autotranslateEndpoints = API.v1
+	.get(
+		'autotranslate.getSupportedLanguages',
+		{
+			authRequired: true,
+			query: isAutotranslateGetSupportedLanguagesParamsGET,
+			response: {
+				200: ajv.compile<{ languages: ISupportedLanguage[] }>({
+					type: 'object',
+					properties: {
+						languages: {
+							type: 'array',
+							items: {
+								type: 'object',
+								properties: {
+									language: { type: 'string' },
+									name: { type: 'string' },
+								},
+								required: ['language', 'name'],
+								additionalProperties: false,
+							},
+						},
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['languages', 'success'],
+					additionalProperties: false,
+				}),
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+			},
+		},
+		async function action() {
 			if (!settings.get('AutoTranslate_Enabled')) {
 				return API.v1.failure('AutoTranslate is disabled.');
 			}
@@ -31,17 +125,26 @@ API.v1.addRoute(
 
 			return API.v1.success({ languages: languages || [] });
 		},
-	},
-);
-
-API.v1.addRoute(
-	'autotranslate.saveSettings',
-	{
-		authRequired: true,
-		validateParams: isAutotranslateSaveSettingsParamsPOST,
-	},
-	{
-		async post() {
+	)
+	.post(
+		'autotranslate.saveSettings',
+		{
+			authRequired: true,
+			body: isAutotranslateSaveSettingsParamsPOST,
+			response: {
+				200: ajv.compile<void>({
+					type: 'object',
+					properties: {
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['success'],
+					additionalProperties: false,
+				}),
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+			},
+		},
+		async function action() {
 			const { roomId, field, value, defaultLanguage } = this.bodyParams;
 			if (!settings.get('AutoTranslate_Enabled')) {
 				return API.v1.failure('AutoTranslate is disabled.');
@@ -70,74 +173,48 @@ API.v1.addRoute(
 
 			return API.v1.success();
 		},
-	},
-);
-
-type AutotranslateTranslateMessageParamsPOST = {
-	messageId: string;
-	targetLanguage?: string;
-};
-
-const AutotranslateTranslateMessageParamsPostSchema = {
-	type: 'object',
-	properties: {
-		messageId: {
-			type: 'string',
+	)
+	.post(
+		'autotranslate.translateMessage',
+		{
+			authRequired: true,
+			body: isAutotranslateTranslateMessageParamsPOST,
+			response: {
+				200: ajv.compile<{ message: IMessage }>({
+					type: 'object',
+					properties: {
+						message: { $ref: '#/components/schemas/IMessage' },
+						success: { type: 'boolean', enum: [true] },
+					},
+					required: ['message', 'success'],
+					additionalProperties: false,
+				}),
+				400: validateBadRequestErrorResponse,
+				401: validateUnauthorizedErrorResponse,
+			},
 		},
-		targetLanguage: {
-			type: 'string',
-			nullable: true,
+		async function action() {
+			const { messageId, targetLanguage } = this.bodyParams;
+			if (!settings.get('AutoTranslate_Enabled')) {
+				return API.v1.failure('AutoTranslate is disabled.');
+			}
+			if (!messageId) {
+				return API.v1.failure('The bodyParam "messageId" is required.');
+			}
+			const message = await Messages.findOneById(messageId);
+			if (!message) {
+				return API.v1.failure('Message not found.');
+			}
+
+			const translatedMessage = await translateMessage(targetLanguage, message);
+
+			if (!translatedMessage) {
+				return API.v1.failure('Failed to translate message.');
+			}
+
+			return API.v1.success({ message: translatedMessage });
 		},
-	},
-	required: ['messageId'],
-	additionalProperties: false,
-};
-
-const isAutotranslateTranslateMessageParamsPOST = ajv.compile<AutotranslateTranslateMessageParamsPOST>(
-	AutotranslateTranslateMessageParamsPostSchema,
-);
-
-const autotranslateEndpoints = API.v1.post(
-	'autotranslate.translateMessage',
-	{
-		authRequired: true,
-		body: isAutotranslateTranslateMessageParamsPOST,
-		response: {
-			200: ajv.compile<{ message: IMessage }>({
-				type: 'object',
-				properties: {
-					message: { $ref: '#/components/schemas/IMessage' },
-					success: { type: 'boolean', enum: [true] },
-				},
-				required: ['message', 'success'],
-				additionalProperties: false,
-			}),
-			400: validateBadRequestErrorResponse,
-			401: validateUnauthorizedErrorResponse,
-		},
-	},
-	async function action() {
-		const { messageId, targetLanguage } = this.bodyParams;
-		if (!settings.get('AutoTranslate_Enabled')) {
-			return API.v1.failure('AutoTranslate is disabled.');
-		}
-		if (!messageId) {
-			return API.v1.failure('The bodyParam "messageId" is required.');
-		}
-		const message = await Messages.findOneById(messageId);
-		if (!message) {
-			return API.v1.failure('Message not found.');
-		}
-
-		const translatedMessage = await translateMessage(targetLanguage, message);
-
-		if (!translatedMessage) {
-			return API.v1.failure('Failed to translate message.');
-		}
-
-		return API.v1.success({ message: translatedMessage });
-	},
-);
+	);
 
 type AutotranslateEndpoints = ExtractRoutesFromAPI<typeof autotranslateEndpoints>;
