--- packages/rest-typings/src/v1/autotranslate/AutotranslateGetSupportedLanguagesParamsGET.ts
@@ -1,20 +1,0 @@
-import { ajv } from '../Ajv';
-
-export type AutotranslateGetSupportedLanguagesParamsGET = {
-	targetLanguage: string;
-};
-
-const AutotranslateGetSupportedLanguagesParamsGETSchema = {
-	type: 'object',
-	properties: {
-		targetLanguage: {
-			type: 'string',
-		},
-	},
-	required: ['targetLanguage'],
-	additionalProperties: false,
-};
-
-export const isAutotranslateGetSupportedLanguagesParamsGET = ajv.compile<AutotranslateGetSupportedLanguagesParamsGET>(
-	AutotranslateGetSupportedLanguagesParamsGETSchema,
-);
